+++ @xen-orchestra/web/src/modules/site/composables/xo-site-tree.composable.ts
@@ -3,13 +3,12 @@ import { useXoPoolCollection } from '@/modules/pool/remote-resources/use-xo-pool
 import type { XoSite } from '@/modules/site/types/xo-site.type.ts'
 import { useXoVmCollection } from '@/modules/vm/remote-resources/use-xo-vm-collection.ts'
 import { XOA_NAME } from '@/shared/constants.ts'
-import type { TreeNodeId } from '@core/packages/tree/types.ts'
+import { useTreeCollapse } from '@core/composables/tree-collapse.composable.ts'
 import { useTreeFilter } from '@core/composables/tree-filter.composable.ts'
 import { defineTree } from '@core/packages/tree/define-tree.ts'
 import { useTree } from '@core/packages/tree/use-tree.ts'
-import { useLocalStorage } from '@vueuse/core'
 import { logicAnd } from '@vueuse/math'
-import { computed, reactive, ref, watch } from 'vue'
+import { computed } from 'vue'
 
 export function useXoSiteTree() {
   const { pools, arePoolsReady } = useXoPoolCollection()
@@ -71,23 +70,7 @@ export function useXoSiteTree() {
     )
   )
 
-  const collapseState = reactive({
-    default: useLocalStorage('site.collapsed', new Set<TreeNodeId>()),
-    filtered: ref(new Set<TreeNodeId>()),
-  })
-
-  watch(filter, () => collapseState.filtered.clear())
-
-  const collapsedIds = computed({
-    get: () => (hasFilter.value ? collapseState.filtered : collapseState.default),
-    set: value => {
-      if (hasFilter.value) {
-        collapseState.filtered = value
-      } else {
-        collapseState.default = value
-      }
-    },
-  })
+  const collapsedIds = useTreeCollapse('site.collapsed', filter, hasFilter)
 
   const { flatNodes, flatNodeIndexById, expandToNode } = useTree(definitions, { collapsedIds })